% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{auszahlungen_sti_korrelation}
\alias{auszahlungen_sti_korrelation}
\title{Korrelation zwischen Auszahlungen und STI}
\usage{
auszahlungen_sti_korrelation(path, sheetnr)
}
\arguments{
\item{path}{Gibt an, wo sich die relevante Tabelle befindet. Wird path nicht
angegeben, so wird die im Package vorhandene Tabelle benutzt.}

\item{sheetnr}{Gibt an, auf welcher Seite der Excel Tabelle die releanten
Daten zu finden sind. Der Parameter muss nur angegeben werden, falls ein
eigener Datensatz verwendet wird.}
}
\value{
Gibt den täglichen Korrelationskoeffizienten zwischen Lohnauszahlungen
und der STI in allen Landkreisen als Zeitreihe in der Form eines Tibbles
zurück. Die Spalte "date" beschreibt das Datum, die Spalte "cor" den Koeffizienten.
}
\description{
Diese Funktion berechnet die Korellation zwischen der STI eines Landkreises
und dem durchschnittlichen Arbeiterentgeld (i.e. ausgezahlter Lohn) des Landkreises
für jeden Tag aus. Anschließend wird der Korrelationskoeffizient als Zeitreihe ausgegeben.
}
\details{
Die Funktion basiert auf Daten des Destatis, bezogen
von https://www.statistikportal.de/de/vgrdl/ergebnisse-kreisebene/einkommen-kreise
am 12.8.2021. Man kann auch eine eigene Tabelle einladen. Hierfür muss der entsprechende
Pfad, als auch die Sheetnumber des relevanten Datensatz angegeben werden.
}
\examples{
auszahlungen_sti_korrelation()
auszahlungen_sti_korrelation("expl/path.xlsx", 11)

\dontrun{auszahlungen_sti_korrelation(sheetnr=3)}
\dontrun{auszahlungen_sti_korrelation(path="custom/path/without/sheetnumber.xlsx")}

}
\concept{Wohlstand und Corona}
