% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_ages.R
\name{calc_covid_mortality}
\alias{calc_covid_mortality}
\title{Time series of COVID-19 associated mortality}
\usage{
calc_covid_mortality(
  ages = "all",
  regions = "Germany",
  from = "2020-01-01",
  to = Sys.Date()
)
}
\arguments{
\item{ages}{A vector of integer numbers that specify the desired age groups.}

\item{regions}{A vector that either consists of strings (the names of German districts or the names of German states) or district ID's.
If this vector contains more than entry, the function calculates the combined mortality of these districts or states.}

\item{from}{A date that specifies the beginning of the time series}

\item{to}{A date that specifies the end of the time series}
}
\value{
A tibble that contains the COVID-19 associated mortality per age group and time period.
}
\description{
\code{calc_covid_mortality()} is used to create a tibble of the COVID-19 mortality per age group and user-definded
time period.
}
\section{Warning}{

These numbers must be handled carefully because of the unknown dark figure.
}

\examples{
calc_covid_mortality(ages=c(10,76,42),regions=c("Bayern","Sachsen"))

calc_covid_mortality(ages="all",regions=8221,from="2020-02-03",to="2021-02-03")

}
\seealso{
Other COVID-19 associated mortality: 
\code{\link{plot_covid_mortality}()}
}
\concept{COVID-19 associated mortality}
