% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk_correlation.R
\name{calc_sti_correlation_of_lks}
\alias{calc_sti_correlation_of_lks}
\title{Incidence Correlation of District Pairs}
\usage{
calc_sti_correlation_of_lks(regions = "all")
}
\arguments{
\item{regions}{A vector of strings with district names or district IDs which should be compared.
If the value is "all", all districts will be compared.}
}
\value{
A tibble with all pairs of given regions, sorted by the ones with the highest correlation first.
}
\description{
This function allows the user to calculate the empirical correlation of district pairs
over the pandemic. The correlation is calculated by looking at the empirical
correlation of the STI of one district with the STI of the other district on a daily basis
}
\section{Warning}{

When calling this function with the paramater "all" (the default paramater),
the results may take a while, because this is a very expensive computation.
}

\examples{
calc_sti_correlation_of_lks(c("Heidelberg", 8222, "1001"))

}
