% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare_system.R
\name{calc_traced_cases}
\alias{calc_traced_cases}
\title{Time Series of Traced Cases over Pandemic}
\usage{
calc_traced_cases(
  ages = "all",
  regions = "Germany",
  from = "2020-01-01",
  to = Sys.Date()
)
}
\description{
\code{calc_traced_cases()} is used to create a time series for the pandemic
which shows the total and relative amount of cases traced by the German healthcare system.
Traced cases here means that the infection date of an infection is known,
as opposed to cases which had a positive test, but could not be traced back to the infection.
}
