% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{east_germany}
\alias{east_germany}
\title{East Germany}
\usage{
east_germany()
}
\value{
A tibble with a "date" and a "mean_sti" column.
}
\description{
Generates an STI time series for the new federal states.
}
\seealso{
Other East-West: 
\code{\link{west_germany}()}
}
\concept{East-West}
