% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitness.R
\name{Fitness}
\alias{Fitness}
\title{Fitness: F(X)}
\usage{
Fitness(inOrg, fun = "max_x")
}
\arguments{
\item{inOrg}{The input organism-object}

\item{fun}{A fitness-evaluation function to be applied to each organism}
}
\value{
fit A numeric vector of the fitness of each organism
}
\description{
Given an organism ran in the Game of Life.
Evaluate it's fitness based on a fitness-function.
}
\examples{
organism_fitness <- Fitness(glider, "max_x")

# There is a wrapper function to apply to `populations` directly
gliders_G0@fitness <- popFitness(gliders_G0)

}
