% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGenerationFirst.R
\name{generateGenerationFirst}
\alias{generateGenerationFirst}
\title{generateGenerationFirst}
\usage{
generateGenerationFirst(gridsize, density, numOfOrganisms, folderName)
}
\arguments{
\item{gridsize}{- vector of length 2: number of rows and number of columns}

\item{density}{- density of active cells}

\item{numOfOrganisms}{- number of organisms}

\item{folderName}{- output folder}
}
\value{
returns population object with generated random organisms
}
\description{
function that generates population of initial organisms for Game of Evolution
}
\examples{
popG0 <- generateGenerationFirst(c(100,100), 0.5, 100, "~/Desktop/test_bgolly/G0")
}
