% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenFiltMask.R
\name{GenFiltMask}
\alias{GenFiltMask}
\title{SGenFiltMask}
\usage{
GenFiltMask(nrow, ncol, N, SD)
}
\arguments{
\item{file_path}{full path of pattern file along with the name}
}
\value{
A FiltMask-class object
}
\description{
GenFiltMask function
Input size of the matrix
}
\details{
Accepts four arguments - nrow, ncol, N, SD
and returns N `FiltMask` objects
}
\examples{


}
