% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\name{get_abs_deaths}
\alias{get_abs_deaths}
\title{Time series of the total number of daily deaths in Germany}
\usage{
get_abs_deaths(years = 2020)
}
\arguments{
\item{years}{A vector of years. The years must be represented as integers. Data is available for the years 2000 up to 2020.}
}
\value{
A tibble that contains the daily death data of the years the user specified.
}
\description{
\code{get_abs_deaths} is used to create a tibble of the total number of deaths during the time span of a year.
It is possible to include more than one year.
}
\examples{
get_abs_deaths(c(2001,2003,2006,2016))

\dontrun{get_abs_deaths(1999)}
#Data is only available for 2000-2020

\dontrun{get_abs_deaths(c("2006"))}
#The years must be integer values
}
