% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{get_accidents_data}
\alias{get_accidents_data}
\title{Traffic accidents of 2019 and 2020}
\usage{
get_accidents_data()
}
\description{
\code{get_accidents_data()} is used to create a tibble of the monthly number of traffic accidents during 2019-2020.
}
\examples{
get_accidents_data()

}
\concept{Traffic accidents}
