% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sti.R
\name{get_data_for}
\alias{get_data_for}
\title{Creating Time Series of Cases, Deaths or Incidence}
\usage{
get_data_for(
  regions,
  ages = "all",
  from = "2020-01-01",
  to = Sys.Date(),
  type = "cases"
)
}
\arguments{
\item{regions}{A vector that either consists of strings (the names of German districts or the names of German states) or district ID's, or just "Germany" as a whole
If this vector has more than one entry, the tibble contains the values of these regions separately.}

\item{ages}{A vector of numbers specifying the desired age groups. The available age groups are "A00-A04","A05-A14","A15-A34","A35-A59","A60-A79" and "A80+".
The numbers in \code{ages} are automatically assigned to the belonging age group and afterwards, the cases and deaths of of these age groups are added up.}

\item{from}{A date that specifies the beginning of the time series}

\item{to}{A date that specifies the end of the time series}

\item{type}{A string specifying which time series is returned. Can be "sti" for the incidence or just "cases" / "deaths"}
}
\value{
A tibble that contains the wanted value for the desired regions during the defined period of time.
}
\description{
\code{get_data_for} is used to create a table of the absolute number of cases, of deaths or of
the 7-day-incidence of a region and a certain
age group. The time period can be defined by the user.
}
\section{Warning}{

When specifying region \strong{and} agegroup, the incidence will not be accurate because
there is no population data for the age groups in each district and it will be estimated
by the age distribution of Germany.
Therefore, it is recommended to specify only one or the other
}

\examples{
get_data_for(ages=c(3,55),regions="Sachsen",from="2021-01-05",to="2021-02-06")

get_data_for(ages=32,regions=c(8221, "Hessen", "Germany"), type="sti")

\dontrun{get_time_series_for(ages=c(12,42),from="2020-05-02",to="2020-05-01"}
#"from" must always be an earlier date than "to"

}
\seealso{
Other time series of cases, deaths or incidence: 
\code{\link{plot_data_for}()}
}
\concept{time series of cases, deaths or incidence}
