% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{get_public_transportation_data}
\alias{get_public_transportation_data}
\title{Kilometers per person in traffic}
\usage{
get_public_transportation_data()
}
\description{
\code{get_public_transportation_data} is used to create a tibble that contains quarterly data of kilometers per person of the years 2019 and 2020.
}
\examples{
get_public_transportation_data

}
\seealso{
Other traffic kilometers: 
\code{\link{plot_public_transportation_with_sti}()}
}
\concept{traffic kilometers}
