% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/god.R
\name{god}
\alias{god}
\title{god}
\usage{
god(I, J, density)
}
\arguments{
\item{I}{number of rows in organism}

\item{J}{number of columns in organism}

\item{density}{Range = [0,1]. Probabilty a given cell is set to 1}
}
\value{
An organism-class object
}
\description{
The spark of all life in the Game of Evolution
aka organism initialization function
}
\details{
Create a binary matrix of size I x J
and return it as an `organism` object
}
\examples{

# Create a 9x9 matrix with 50\% "on" density
god(I = 9, J = 9, density = 0.5)

}
