% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{income_sti_correlation}
\alias{income_sti_correlation}
\title{Correlation between income and STI}
\usage{
income_sti_correlation(path, sheetnr)
}
\arguments{
\item{path}{Specifies the path to a giben excel sheet. If left empty the
function defaults to the data provided by the pachage.}

\item{sheetnr}{Specifies the sheet of the excel table, where the relevant data can
be found. It only has to be specified, if custom data is used.}
}
\value{
The function returns a tibble carring a "date" column, and a "cor" column.
The "cor" column describes the correlation for a given date.
}
\description{
This function calculates the correlation between average income and seven day
incidence in all administrative regions of germany on a dayly basis.
}
\details{
The funciton is based on data by Destatis aquired via
https://www.statistikportal.de/de/vgrdl/ergebnisse-kreisebene/einkommen-kreise
on the 12.8.2021. The user can load his own income data into the function, as
described in the parameters.
}
\examples{
income_sti_correlation()
income_sti_correlation("expl/path.xlsx", 11)

\dontrun{income_sti_correlation(sheetnr=3)}
\dontrun{income_sti_correlation(path="custom/path/without/sheetnumber.xlsx")}

}
