% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\docType{class}
\name{lineage-class}
\alias{lineage-class}
\alias{lineage}
\title{lineage - class definition}
\arguments{
\item{populations}{An ordered list of populations at geneartions Gi, Gi+j, Gi+2j, ...}

\item{G}{An ordered numerical vector of generation numbers for each population c(Gi, Gi+j,...)}
}
\value{
lineage
}
\description{
The population for a geneartion (G_i) is processed
through the Game of Evolution to yield the next generation (G_i+1).
}
\details{
Every jth generation, the `population` will be written to the `lineage`
object which can be used for tracking changes and plotting
}
\examples{


}
