% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_load.R
\name{manual_load_rki_data}
\alias{manual_load_rki_data}
\title{Manually load RKI data}
\usage{
manual_load_rki_data(path)
}
\arguments{
\item{path}{A path to a file which is to be loaded in. The path has to be in parentheses.}
}
\description{
This function manually laods an RKI file into the package. The user has to provide a valid
path to a valid RKI file. If a wrong file is loaded, the package cannot build its internal
data. Thus we recommend using the automatic \code{update_rki_data()} function.
}
\details{
Please be aware of the front-slash/back-slash conventions of your operating system.
}
\examples{
manual_load_rki_data("~/Downloads/RKI.csv")
\dontrun{manual_load_rki_data()}
\dontrun{manual_load_rki_data(~/Downloads/RKI.csv)}

}
\seealso{
Other update: 
\code{\link{update_rki_data}()},
\code{\link{update_vac_data}()},
\code{\link{update_voc_data}()}
}
\concept{manually load}
\concept{update}
