% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2rle.R
\name{mat2rle}
\alias{mat2rle}
\title{mat2rle}
\usage{
mat2rle(mat, fileName)
}
\arguments{
\item{file_path}{full path of pattern file along with the name}
}
\value{
An organism-class object
}
\description{
pattern reading function
read a .rle pattern file and convert it to I x J matrix
}
\details{
Accepts a .rle pattern file
and returns a `organism` object
}
\examples{

# Create a 9x9 matrix with 50\% "on" density
rle2mat(file_path)

}
