% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2rle_2.R
\name{mat2rle_2}
\alias{mat2rle_2}
\title{mat2rle}
\usage{
mat2rle_2(organism_l, fileName)
}
\arguments{
\item{organism_l}{an organism object}
}
\description{
pattern writing function
convert a binary I x J matrix to a .rle file readable by golly
}
\details{
Accepts a binary matrix
and writes a organism.rle file as output
}
\examples{

mat2rle(organism_l)

}
