% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{Mutation}
\alias{Mutation}
\title{Mutation}
\usage{
Mutation(org_bin_matr, md = c("constant", "guassian", "poisson"),
  mut_rate_avg, mut_rate_sd = NULL)
}
\arguments{
\item{file_path}{full path of pattern file along with the name}
}
\value{
An organism-class object
}
\description{
organism mutation function
Input a binary organism matrix with selected mutation distribution selection
}
\details{
Accepts a binary matrix
and returns a `organism` object
}
\examples{




}
