% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weekday_Distribution.R
\name{office_case_distribution}
\alias{office_case_distribution}
\title{Comparison of offices from different federal states regarding the number of reported cases on the different weekdays.}
\usage{
office_case_distribution(
  bundesland_1 = "Baden-Württemberg",
  bundesland_2 = "Schleswig-Holstein"
)
}
\arguments{
\item{bundesland_1}{A string of the desired federal state.}

\item{bundesland_2}{A string of the other desired federal state for the comparison.}
}
\value{
A plot that shows the standardized amount of cases that get reported per day in the given federal states
and the standardized amount of all reported cases per day.
}
\description{
\code{office_case_distribution()} is used to create a plot showing the mean number of reported cases in Germany regarding the specific weekdays
as well as the number of reported cases regarding the specific weekdays of the two given federal states. To see how fast or slow offices of different
federal states are reporting cases and how they differ to the mean number of reported cases overall.
All values are standardized.
}
\examples{
office_case_distribution("Hamburg","Niedersachsen")
}
