% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\docType{class}
\name{organism-class}
\alias{organism-class}
\alias{organism}
\title{organism - class definition}
\arguments{
\item{cells}{an IxJ binary matrix to be stored as an organism}

\item{shift}{a Vector of c(X,Y) coordinate offsets to center the cells matrix [c(0,0)]}
}
\value{
organism
}
\description{
In the Game of Evolution, an organism is an I x J binary matrix.
Each ON state in this matrix is a 'living cell' in Game of Life
Each OFF state in this matrix is a 'dead cell' in Game of Life
}
\details{
To account for
}
\examples{
# Example of a glider encoded as logical matrix
glider_logical <- matrix( data = c(F,T,F,
                                   F,F,T,
                                   T,T,T), nrow = 3, byrow = T)

# Example of a glider encoded as an organism
glider <- organism(cells = glider_logical)

}
