% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{ost_deutschland}
\alias{ost_deutschland}
\title{Ost Deutschland}
\usage{
ost_deutschland()
}
\value{
Ein Tibble mit einer Spalte "date" und einer Spalte "mean_sti".
}
\description{
Generiert die STI time series für die neuen Bundesländer.
}
\seealso{
Other Ost-West: 
\code{\link{west_deutschland}()}
}
\concept{Ost-West}
