% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P53_sequences}
\alias{P53_sequences}
\title{25 sequences from P53 protein}
\format{
A list of 216 lines from 25 sequences:
\describe{
  \item{25 sequences}{25 protein sequences.}
}
}
\source{
<UniProt>
}
\usage{
P53_sequences
}
\description{
a list of 25 P53 protein sequences.
}
\keyword{datasets}
