% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{payments_sti_correlation}
\alias{payments_sti_correlation}
\title{Correlation between payouts and STI}
\usage{
payments_sti_correlation(path, sheetnr)
}
\arguments{
\item{path}{Specifies where the relevant table is located. If path is not
specified, the table available in the package will be used.}

\item{sheetnr}{Specifies on which page of the Excel table the relevant
data can be found. The parameter must only be specified if an
own dataset is used.}
}
\value{
Gives the daily correlation coefficient between wage
payments and the
and the STI in all counties as a time series in the form of a tibble
The "date" column describes the date, and the "cor" column describes the coefficient.
}
\description{
This function calculates the correlation between the STI of a county
and the average worker compensation (i.e., wages paid out) of the county
for each day. It then outputs the correlation coefficient as a time series.
}
\details{
The function is based on data from Destatis, sourced
from https://www.statistikportal.de/de/vgrdl/ergebnisse-kreisebene/einkommen-kreise
on 12.8.2021. You can also load your own table. For this the corresponding
path, as well as the sheetnumber of the relevant dataset.
}
\examples{
payments_sti_correlation()
payments_sti_correlation("expl/path.xlsx", 11)

\dontrun{payments_sti_correlation(sheetnr=3)}
\dontrun{payments_sti_correlation(path="custom/path/without/sheetnumber.xlsx")}

}
\concept{wealth and corona}
