% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_population_composition.R
\name{Pl_pc}
\alias{Pl_pc}
\title{Plot population compositon: Pl_pc()}
\usage{
Pl_pc(inANS)
}
\arguments{
\item{ancestry}{An ancestry object for the population}
}
\value{
A plottable composiont graph
}
\description{
Given an ancestry object generate a population composition graph
#Requires ggplot2
}
\examples{
pl <- Pl_pc(ancestry)
plot(pl)


}
