% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\name{plot_abs_deaths}
\alias{plot_abs_deaths}
\title{Plotting the total number of daily deaths in Germany}
\usage{
plot_abs_deaths(years = 2020, smoothing = 0)
}
\arguments{
\item{years}{A vector of years. The years must be represented as integers. Data is available for the years 2000 up to 2020.}

\item{smoothing}{A positive integer that defines the window size of the moving average. Thus, the plot will be smoother
the higher 'smoothing' is chosen. The default setting is 'no smoothing'.}
}
\value{
A plot of the daily death data of the years the user specified.
}
\description{
\code{plot_abs_deaths} is used to create a plot of the total number of deaths during the time span of a year.
It is possible to include more than one year.
}
\examples{
get_abs_deaths(c(2001,2003,2006,2016),smoothing=2)

\dontrun{get_abs_deaths(1999)}
#Data is only available for 2000-2020.

\dontrun{get_abs_deaths(c("2006"),smoothing=-2.4)}
#Years must be a vector of integer values and smoothing must be a positive integer.
}
