% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{plot_accidents_with_sti}
\alias{plot_accidents_with_sti}
\title{Comparison between traffic accidents and 7-day-incidence}
\usage{
plot_accidents_with_sti()
}
\description{
\code{plot_accidents_with_sti()} is used to create a plot of the monthly number of traffic accidents in comparison with
the 7-day-incidence in Germany. Obviously, the number of accidents drastically sank when COVID-19 reached Germany.
}
\examples{
plot_accidents_with_sti()
}
