% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_ages.R
\name{plot_covid_mortality}
\alias{plot_covid_mortality}
\title{Plotting the COVID-19 associated mortality}
\usage{
plot_covid_mortality(
  ages = "all",
  regions = "Germany",
  from = "2020-01-01",
  to = Sys.Date(),
  smoothing = 0
)
}
\arguments{
\item{ages}{A vector of integer numbers that specify the desired age groups.}

\item{regions}{A vector that either consists of strings (the names of German districts or the names of German states) or district ID's.
If this vector contains more than entry, the function plots the combined mortality of these districts or states.}

\item{from}{A date that specifies the beginning of the time series}

\item{to}{A date that specifies the end of the time series}

\item{smoothing}{A positive integer that defines the window size of the moving average. Thus, the plot will be smoother
the higher 'smoothing' is chosen. The default setting is 'no smoothing'.}
}
\value{
A plot of the COVID-19 associated mortality per age group and time period.
}
\description{
\code{plot_covid_mortality()} is used to create a plot of the COVID-19 mortality per age group and user-definded
time period.
}
\section{Warning}{

These numbers must be handled carefully because of the unknown dark figure.
}

\examples{
plot_covid_mortality(ages=c(10,76,42),regions=c("Bayern","Sachsen"))

plot_covid_mortality(ages="all",regions=8221,from="2020-02-03",to="2021-02-03")
}
