% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare_system.R
\name{plot_distribution_report_diff}
\alias{plot_distribution_report_diff}
\title{Plot Time Difference Between Infection and Report of Cases}
\usage{
plot_distribution_report_diff(
  ages = "all",
  regions = "Germany",
  from = "2020-01-01",
  to = Sys.Date(),
  cut = 30
)
}
\description{
Plots the data of \code{calc_distribution_report_diff}. See \code{\link{calc_distribution_report_diff}} for an explanation of the data
}
