% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\name{plot_excess_mortality}
\alias{plot_excess_mortality}
\title{Plot of excess mortality}
\usage{
plot_excess_mortality(
  excess_year = 2020,
  average_years = c(2016, 2017, 2018, 2019),
  smoothing = 0
)
}
\arguments{
\item{excess_year}{An integer that stands for the year of which the total deaths should be plotted against the average.}

\item{average_years}{A vector of integers that represent the years of which the function should calculate the average deaths.}

\item{smoothing}{A positive integer that defines the window size of the moving average. Thus, the plot will be smoother
the higher 'smoothing' is chosen. The default setting is 'no smoothing'.}
}
\value{
A plot that shows the total number of deaths in comparison to a user-defined average.
}
\description{
\code{plot_excess_mortality} creates a plot of the total number of daily deaths of one year in contrast to the
average values of other years. Data is available for the years 2000-2020.
}
\section{Warning}{

Due to missing data for February 29th in non-leap years, the smoothing function increases the gap in the plot the higher the
smoothing factor is chosen.
}

\examples{
plot_excess_mortality(2020,c(2015,2016,2017,2018,2019),smoothing=1)

\dontrun{plot_excess_mortality(excess_year="2020",average_years=c(1999,2017),smoothing=100)}
#"excess_year" and "average_years" should be an integer. Only the years 2000-2020 can be considered.
#Furthermore, it's not recommended to choose a smoothing factor > 5 due to the gap in the plot described in section "Warning".
}
