% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_density.R
\name{plot_pop_density_with_linear_model}
\alias{plot_pop_density_with_linear_model}
\title{Plot Correlation Between Population Density and STI}
\usage{
plot_pop_density_with_linear_model(regions = "all")
}
\value{
A plot which contains the average STI over the pandemic with the population density of the district
}
\description{
Utilizes a scatter plot to display the data provided by \code{\link{get_pop_density_with_sti}}.
The plot has the population density on the x axis and the average STI on the y axis.
Additionally, a linear model is fit to the data, to visualize a possible correlation.
}
\section{Warning}{

When calling this function with the paramater "all" (the default paramater),
the results may take a while, because this is a very expensive computation.
}

\examples{
get_pop_density_with_sti(regions=c(8221, "Mannheim"))
}
