% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{plot_public_transportation_with_sti}
\alias{plot_public_transportation_with_sti}
\title{Comparison between traffic kilometers per person and 7-day-incidence}
\usage{
plot_public_transportation_with_sti()
}
\description{
\code{plot_public_transportation_with_sti} is used to create a plot of the person-kilometers in comparison with
the 7-day-incidence in Germany. Obviously, the number of kilometers drastically sank when COVID-19 reached Germany.
}
\examples{
plot_public_transportation_with_sti

}
\seealso{
Other traffic kilometers: 
\code{\link{get_public_transportation_data}()}
}
\concept{traffic kilometers}
