% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.R
\name{plot_total_mortality}
\alias{plot_total_mortality}
\title{Plotting the yearly mortality per age group}
\usage{
plot_total_mortality(age = "A80+")
}
\arguments{
\item{age}{A string that indicates the desired age group.
Possible age groups are "A00-A34", "A35-A59", "A60-A79" and "A80+".}
}
\value{
A plot that shows the yearly mortality per age group.
}
\description{
\code{plot_total_mortality} is used to create a plot of the yearly mortality per age group since 2014.
}
\examples{
plot_total_mortality(age="A35-A59")

\dontrun{plot_total_mortality(age=65)}
#The age group must be a string. Allowed strings are "A00-A34","A35-A59","A60-79" and "A80+".

}
\seealso{
Other Yearly mortality per age group: 
\code{\link{get_total_mortality}()}
}
\concept{Yearly mortality per age group}
