% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{plot_variant_data}
\alias{plot_variant_data}
\title{Plot variant data}
\usage{
plot_variant_data(interpolation = "none")
}
\arguments{
\item{interpolation}{Specifies whether the voc data should be interpolated.
Possible options include "linear" or "none". Linear does a linear spline
interpolation. None is the default.}
}
\value{
Returns a ggplot2 plot.
}
\description{
A function that plots share of variant data over time.
}
\examples{
plot_variant_data(), plot_variant_data("linear")

}
\seealso{
Other Variants: 
\code{\link{plot_variants}()},
\code{\link{variant_case_r_value}()},
\code{\link{variant_case_time_series}()},
\code{\link{variant_sti_time_series}()}
}
\concept{Variants}
