% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{plot_variants}
\alias{plot_variants}
\title{Plot variant share}
\usage{
plot_variants(interpolation = "linear", type = "cases")
}
\arguments{
\item{interpolation}{Specifies whether the voc data should be interpolated.
Possible options include "linear" or "none". Linear does a linear spline
interpolation. None is the default.}

\item{type}{String, specifies which value for the variants should be plotted
and how it should be displayed}
}
\value{
Returns a ggplot2 plot.
}
\description{
A function that plots variant data over time.
}
\details{
The function can ether plot the case time series of different vocs, or the
sti time series of different vocs.
}
\examples{
plot_variants()
plot_variants(interpolation="linear", type="sti")
\dontrun{plot_variants(interpolation=T)}

}
\seealso{
Other Variants: 
\code{\link{variant_case_time_series}()},
\code{\link{variant_sti_time_series}()}
}
\concept{Variants}
