% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFitness.R
\name{plotFitness}
\alias{plotFitness}
\title{plotFitness}
\usage{
plotFitness(si_life, no_organisms)
}
\arguments{
\item{file_path}{}
}
\value{
A list of organism-class object
}
\description{
plot the fitness of organism function
Input Lineage object
}
\examples{




}
