% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popMutation.R
\name{popMutation}
\alias{popMutation}
\title{mutation}
\usage{
popMutation(org_population, md = c("constant"), mut_rate_avg = 0.1)
}
\arguments{
\item{org_population}{tmp}

\item{md}{tmp}

\item{mut_rate_avg}{tmp}
}
\value{
A list of organism-class object
}
\description{
organism mutation_pop function
Input population of organisms to undergo mutation
}
\details{
Accepts list of binary matrix
and returns a list of `organism`
}
\examples{




}
