% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{population-class}
\alias{population-class}
\alias{population}
\title{population - class definition}
\arguments{
\item{population}{Class definition for a list of organisms.}
}
\value{
None
}
\description{
For each geneartion (G_i) in the Game of Evolution
There is a collection of `organism` objects which make up a population
which are competing with one another to survive to the next generation.
}
\details{
This is the class definition of a list of organisms and their associated
fitness values
}
\examples{

}
