% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_ancestry.R
\name{rec_anc}
\alias{rec_anc}
\title{Record Ancestry: rec_anc}
\usage{
rec_anc(inPop, rec)
}
\arguments{
\item{inPop}{The input population-object}

\item{rec}{The current ancestry record (optional)}
}
\value{
A record of current ancestry data for the poulation
}
\description{
Records the ancestry of a population for later use
}
\examples{
inital_record <- rec_anc(inPop)
group_record <- rec_anc(inPop,rec)


}
