% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolution-package.R
\docType{package}
\name{revolution-package}
\alias{revolution}
\alias{revolution-package}
\title{revolution: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Frag mich nicht nach Details
}
\section{Hallelujah}{
NA
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{RONA}
