% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2mat.R
\name{rle2mat}
\alias{rle2mat}
\title{rle2mat}
\usage{
rle2mat(bgolly_object)
}
\arguments{
\item{file_path}{full path of pattern file along with the name}
}
\value{
An organism-class object
}
\description{
pattern reading function
read a .rle pattern file and convert it to I x J matrix
}
\details{
Accepts a .rle pattern file
and returns a `organism` object
}
\examples{

rle2mat(file_path)

}
