% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rRNA_sequences}
\alias{rRNA_sequences}
\title{more than 1000 rRNA sequences that have been filtered and cleaned. The results are all complete sequences and not included of N.}
\format{
A list of 216 lines from 25 sequences:
\describe{
  \item{more than 1000 sequencs}{filtered 16S rRNA sequences.}
}
}
\source{
<NCBI/nucleotide>
}
\usage{
rRNA_sequences
}
\description{
a list of more than 1000 rRNA sequences.
}
\keyword{datasets}
