% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rRNA_5S_sequences}
\alias{rRNA_5S_sequences}
\title{more than 11000 5S rRNA sequences from RNAcentral database. The results are all complete sequences and not included of N.}
\format{
A list of 11000 lines from 25 sequences:
\describe{
  \item{more than 11000 sequencs}{5S rRNA sequences.}
}
}
\source{
<RNAcentral: https://rnacentral.org/>
}
\usage{
rRNA_5S_sequences
}
\description{
a list of more than 11000 5S rRNA sequences.
}
\keyword{datasets}
