% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selection.R
\name{Selection}
\alias{Selection}
\title{Selection: Sel(X)}
\usage{
Selection(inPop, por = 0.1)
}
\arguments{
\item{inPop}{The input organism-object}

\item{por}{The portion of population to keep for gen n+1 (Default = 0.1)}
}
\value{
A population of the same size generated from the top portion
}
\description{
Given a population with recorded fitness.
Select best Y portion and repopulate.
}
\examples{
Gn+1 <- Selection(Gn, "surviving_portion")


}
