\name{seq_ds_matrix}
\alias{seq_ds_matrix}
\title{seq_ds_matrix is for getting sequences from a file and give a matrix of 10 parts of FASTA sequences.}
\usage{
seq_ds_matrix(exampleData)
}
\description{
Output will be all parts of FASTA sequences.
}
\examples{
library(REvolution)
data(P53_sequences)
seq_ds_matrix(P53_sequences)
}
