% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_rki_data}
\alias{update_rki_data}
\title{Update RKI data}
\usage{
update_rki_data(method)
}
\arguments{
\item{method}{Specifies via which method the download should happen.
Possible options include "auto", "internal", "libcurl", "wget" and "curl".
If none is chosen, the function defaults to "auto". From experience the
option "auto" works well on Wndows/OSX. For some Linux distributions it's
recommended to choose "wget".}
}
\description{
Loads the newest RKI Covid-19 data via the RKI API, and loads it into the package.
}
\examples{
update_rki_data()
update_rki_data(method="wget")
\dontrun{update_rki_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{manual_load_rki_data}()},
\code{\link{update_vac_data}()},
\code{\link{update_voc_data}()}
}
\concept{update}
