% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_vac_data}
\alias{update_vac_data}
\title{Update Vaccination data}
\usage{
update_vac_data(method)
}
\arguments{
\item{method}{Spezifiziert 체ber welche Methode der Download
passieren soll. Zur Auswahl stehen "auto", "internal", "libcurl",
"wget" und "curl". Wird dieser Parameter nicht spezifiziert, so
wird "auto" gew채lt.}
}
\description{
L채d die aktuellen RKI VOC Daten 체ber die API in das Package.
}
\examples{
update_vac_data()
update_vac_data(method="wget")
\dontrun{update_vac_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{update_rki_data}()},
\code{\link{update_voc_data}()}
}
\concept{update}
