% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_voc_data}
\alias{update_voc_data}
\title{Update Variants of Concern data}
\usage{
update_voc_data(method)
}
\arguments{
\item{method}{Spezifiziert über welche Methode der Download
passieren soll. Zur Auswahl stehen "auto", "internal", "libcurl",
"wget" und "curl". Wird dieser Parameter nicht spezifiziert, so
wird "auto" gewält.}
}
\description{
Läd die aktuellen RKI VOC Daten über die API in das Package. Da das
RKI die zugehörige API öfters verändert, kann es sein, dass diese
Funktion irgentwann nicht mehr funktioniert.
}
\examples{
update_voc_data()
update_voc_data(method="wget")
\dontrun{update_voc_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{update_rki_data}()},
\code{\link{update_vac_data}()}
}
\concept{update}
