% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variant_case_r_value}
\alias{variant_case_r_value}
\title{Variant R value}
\usage{
variant_case_r_value()
}
\value{
A tibble that contains the share of cases with alpha, beta, gamma, delta variant and the according reproduction number (variant specific).
}
\description{
\code{variant_case_R_value()} is used to calculate the R value of different variants of COVID-19 (e.g. alpha, beta, gamma, delta).
The reproduction number (R value) describes how many people an infected person infects on average.
}
\examples{
variant_case_R_value()

}
