% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variant_case_time_series}
\alias{variant_case_time_series}
\title{\code{variant_case_R_value()} is used to calculate the R value of different variants of COVID-19 (e.g. alpha, beta, gamma, delta).
The reproduction number (R value) describes how many people an infected person infects on average.}
\usage{
variant_case_time_series(
  update_data = F,
  interpolation = "none",
  format_long = F
)
}
\arguments{
\item{interpolation}{A string which defines the form of the interpolation. Currently the only option is setting it to "linear" interpolation.
Default value is set to "none"}

\item{tablepath}{Path to "Variance of Cases" table
Default value is set to "system.file("extdata", "VOC_VOI_Tabelle.xlsx",package= "revolution")"}
}
\value{
A tibble that contains the share of cases with alpha, beta, gamma, delta and the according reproduction number (variant specific)

A tibble that contains the share of cases with alpha, beta, gamma, delta variant and if interpolation is set to "linear"
the according interpolation values
}
\description{
\code{variant_case_R_value()} is used to calculate the R value of different variants of COVID-19 (e.g. alpha, beta, gamma, delta).
The reproduction number (R value) describes how many people an infected person infects on average.
}
\examples{
variant_case_R_value()

\code{building_variant_data()} is used to calculate the interpolation values based on the relative share of the specific variant cases.

building_variant_data("linear")
}
