% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variant_sti_time_series}
\alias{variant_sti_time_series}
\title{Variant STI time series}
\usage{
variant_sti_time_series(interpolation = "linear", format_long = F)
}
\arguments{
\item{interpolation}{Specifies whether the voc data should be interpolated. Options
include "linear" and "none". The default in "none", i.e. no interpolation.}

\item{format_long, }{boolean Specifies whether a long format should be used.
A long format pivots the data. The default is F}
}
\value{
A tibble carrying a "date" column, and a column for each variant STI for
that day.
}
\description{
This function takes the weekly variant of concern data provided by the RKI and
approximates the STI time series of the different variants.
}
\details{
It assumes, that the proportions of the variants in the actual cases is the same
as the proportion in the sampled viral matter. Since the RKI only provides the
variant shares on a weekly bases, the function can interpolate the data, using a linear
spline interpolation.
}
\examples{
variant_sti_time_series()
variant_sti_time_series(interpolation="linear")

\dontrun{variant_sti_time_series(interpolation=T)}
}
\seealso{
Other Variants: 
\code{\link{plot_variant_data}()},
\code{\link{plot_variants}()},
\code{\link{variant_case_r_value}()},
\code{\link{variant_case_time_series}()}
}
\concept{Variants}
