% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{west_deutschland}
\alias{west_deutschland}
\title{West Deutschland}
\usage{
west_deutschland()
}
\value{
Ein Tibble mit einer Spalte "date" und einer Spalte "mean_sti".
}
\description{
Generiert die STI time series für die alten Bundesländer.
}
\seealso{
Other Ost-West: 
\code{\link{ost_deutschland}()}
}
\concept{Ost-West}
