% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_sti_correlation.R
\name{west_germany}
\alias{west_germany}
\title{West Germany}
\usage{
west_germany()
}
\value{
A tibble with a "date" and a "mean_sti" column.
}
\description{
Generates an STI time series for the old federal states.
}
\seealso{
Other East-West: 
\code{\link{east_germany}()}
}
\concept{East-West}
