% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2rle.R
\name{mat2rle}
\alias{mat2rle}
\title{mat2rle}
\usage{
mat2rle(organism, fileName)
}
\arguments{
\item{organism}{object of class organism}

\item{fileName}{name of the new file with the pattern}
}
\value{
None. Saves the organism matrix into file
}
\description{
function that writes the organism cell matrix with golly pattern into rle file
}
\details{
Accepts an object of class matrix and file name
and saves the matrix in the rle file format (with the given gile name)
}
\examples{

# Save glider organism object into rle file named CoolOrganism.rle

glider_logical <- matrix( data = c(F,T,F,
F,F,T,
T,T,T), nrow = 3, byrow = T)

glider <- organism(cells = glider_logical)

mat2rle(glider, "CoolOrganism.rle")

}
