% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{Mutation}
\alias{Mutation}
\title{Mutation}
\usage{
Mutation(org_bin_matr, md = c("constant"), mut_rate_avg = 0.1)
}
\arguments{
\item{org_bin_matr}{tmp}

\item{md}{tmp}

\item{mut_rate_avg}{tmp}
}
\value{
An organism-class object
}
\description{
organism mutation function
Input a binary organism matrix with selected mutation distribution selection
}
\details{
Accepts a binary matrix
and returns a `organism` object
}
\examples{




}
