% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sti.R
\name{plot_for_agegroups}
\alias{plot_for_agegroups}
\title{Plotting a Comparison of Age Groups over the Pandemic}
\usage{
plot_for_agegroups(type = "cases")
}
\arguments{
\item{type}{A string which can be "cases", "deats" or "sti".}
}
\description{
\code{plot_for_agegroups} is used to generate a plot that can compare the case/death numbers
and incidence age groups. When comparing the cases or deaths, it returns a plot
which shows the share of cases/deaths for each age group.
}
\examples{
plot_for_agegroups(type="sti")
}
