% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_pop.R
\name{popMutation}
\alias{popMutation}
\title{mutation}
\usage{
popMutation(org_population, md = c("constant", "guassian", "poisson"),
  mut_rate_avg, mut_rate_sd = NULL)
}
\arguments{
\item{file_path}{}
}
\value{
A list of organism-class object
}
\description{
organism mutation_pop function
Input population of organisms to undergo mutation
}
\details{
Accepts list of binary matrix
and returns a list of `organism`
}
\examples{




}
