% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{population-class}
\alias{population-class}
\alias{population}
\title{population - class definition}
\arguments{
\item{organisms}{an ordered List of organism-objects}

\item{fitness}{an ordered Vector of numerical fitness-evaluation values for each organism in population. [NA]}
}
\value{
population
}
\description{
For each geneartion (G_i) in the Game of Evolution
There is a collection of `organism` objects which make up a population
which are competing with one another to survive to the next generation.
}
\details{
This is the class definition of a list of organisms and their associated
fitness values
}
\examples{

# The population at g0 is glider and inverse_glider organisms
# Fitness and maternal line are initially undefined
gliders_G0 <- population( organisms = list( glider, glider_inv), fitness = c(NA,NA) , maternal_line = c(NA,NA) )

}
