% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolution-package.R
\docType{package}
\name{revolution-package}
\alias{revolution}
\alias{revolution-package}
\title{revolution: Analysis of COVID-19 Data in Germany}
\description{
The package "revolution" provides various tools of data analysis to inspect
    the course and impact of the COVID-19 pandemic in Germany over the span of more
    than a year. The package offers the possibility to retrieve the newest data directly
    from the German federal agency for disease control (RKI). It can be used to look at
    the trend of infections and deaths throughout
    the crisis, as well as analyze the circulation of the virus in certain demographic
    groups and compare the vulnerability to the virus of those groups. Furthermore,
    it can display the spread of the different virus mutations and the progress of the
    vaccination campaigns, starting at the end of 2020.
}
\details{
The core functions of the package include \code{get_data_for()},
\code{get_sti_series_for()}, \code{get_vaccination_data()}, \code{variant_sti_time_series()}
and \code{plot_excess_mortality()} with their respective plot functions for
presentiveness.

The package provides multiple different functions: For one, the user can access a
lot of prepared data (eg. in \code{plot_for_agegroups()}
or \code{plot_accidents_with_sti()}), which can be used for further analysis.
On the other hand,
}
\references{
RKI Data: \url{https://npgeo-corona-npgeo-de.hub.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0/about}

Vaccination Data: \url{https://impfdashboard.de/daten}

Variants Data: \url{https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/Daten/VOC_VOI_Tabelle.xlsx?__blob=publicationFile}

Every other dataset in this package is not up to date and comes from the
"Statistisches Bundesamt": \url{https://www.destatis.de/DE/Home/_inhalt.html}.
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Nico Bruder
  \item Lucienne-Sophie Marmé
  \item Viktor Stein von Kamienski
  \item Philipp Hilmar Tepel
}

}
\keyword{internal}
