% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_rki_data}
\alias{update_rki_data}
\title{Update RKI data}
\usage{
update_rki_data(method)
}
\arguments{
\item{method}{Spezifiziert über welche Methode der Download
passieren soll. Zur Auswahl stehen "auto", "internal", "libcurl",
"wget" und "curl". Wird dieser Parameter nicht spezifiziert, so
wird "auto" gewält. Erfahrungsweise funktioniert dies auf Windows/
OSX recht gut, auf Linux wird "wget" empfohlen, da die Datei für "curl"
zu groß ist.}
}
\description{
Läd die aktuellen RKI Daten über die RKI API in das Package
}
\examples{
update_rki_data()
update_rki_data(method="wget")
\dontrun{update_rki_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{update_vac_data}()},
\code{\link{update_voc_data}()}
}
\concept{update}
