% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_vac_data}
\alias{update_vac_data}
\title{Update Vaccination data}
\usage{
update_vac_data(method)
}
\arguments{
\item{method}{Specifies via which method the download should happen.
Possible options include "auto", "internal", "libcurl", "wget" and "curl".
If none is chosen, the function defaults to "auto".}
}
\description{
Loads the newest Vaccination Data via the RKI API and loads it into
the package.
}
\examples{
update_vac_data()
update_vac_data(method="wget")
\dontrun{update_vac_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{update_rki_data}()},
\code{\link{update_voc_data}()}
}
\concept{update}
