% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_voc_data}
\alias{update_voc_data}
\title{Update Variants of Concern data}
\usage{
update_voc_data(method)
}
\arguments{
\item{method}{Specifies via which method the download should happen.
Possible options include "auto", "internal", "libcurl", "wget" and "curl".
If none is chosen, the function defaults to "auto".}
}
\description{
This function gets the newest Variants of Concern data via the RKI API
and loads it into the package.
}
\examples{
update_voc_data()
update_voc_data(method="wget")
\dontrun{update_voc_data(method=wget)}

}
\seealso{
Other update: 
\code{\link{manual_load_rki_data}()},
\code{\link{update_rki_data}()},
\code{\link{update_vac_data}()}
}
\concept{update}
